+++ packages/client/package.json
@@ -12,6 +12,7 @@
     "@testing-library/user-event": "^7.1.2",
     "classnames": "^2.2.6",
     "firebase": "^7.15.5",
+    "formik": "^2.1.5",
     "i18next": "^19.5.4",
     "i18next-browser-languagedetector": "^5.0.0",
     "i18next-http-backend": "^1.0.16",
@@ -23,9 +24,10 @@
     "react-i18next": "^11.7.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.0",
+    "react-test-renderer": "^16.13.1",
     "react-transition-group": "^4.4.1",
     "styled-components": "^5.1.1",
-    "react-test-renderer": "^16.13.1"
+    "yup": "^0.29.3"
   },
   "scripts": {
     "start": "react-scripts start",
